Rework vehicle state getting to be a one-shot operation
using the data_request method forces us to use one RPC per request
which is wasteful and makes us to skip different request types
at the same time, so while we are driving e.g. climate data is
not being updated.
Using vehicle_date request allows us to get all states with a single
request which should be a lot less overhead for us and for Tesla.

the is_asleep still needs to be reworked so we don't create new
connection every time or at least not every time when we are not asleep.

This also adds some logic to keep polling at full speed when the
gear shift is in D or R since even if we are momentarily stopped,
say at a traffic light, we should not really start the slower
polling and potentially miss when we start moving again.

Additionally it remains to be seen if we really can poll this frequently
at 1 request per second or if the car data is fed less frequently to us
for this we really need to look at the timestamps returned and
if the frequence is more rare - adjust?