Keep polling if the car is active even when api commanded to stop
Since the api might be disabled by bt disconnection, we better
keep polling if the car is driving or charging.
Moe conditions could be added later.

If there is a valid usecase, we might want to add a hard stop that
would stop no matter what, but I don't see why for now.