Implement oneshot request to scrape once
This is handful for when you have other external triggers
to indicate car started doing something interesting or
perhaps a one-time run at like midnight to have a definite
reading for say battery-degradation data every day even on
a long-term parked car.
this WILL wake the car if it's sleeping.

Do not set disabledsince outside of post-API since even if we have an
override running - that does not change the time when the scraping
was requested to be disabled.